Refactor : Moved business logic to another component
